Add indirect test for skip height through ancestor search
Add a unit test that indirectly tests the `GetSkipHeight()` function,
by doing checks on the performance of ancestor search that relies on
skip heights. Random searches are performed on a long chain, and the
effective number of steps are counted, and asserted that are much less
than the distance to the ancestor. For boundary value a
log(distance)-based formula is used (the formula is of the form
$A*log(y−x) - B$, with empirically established A and B values).

Co-authored-by: Lőrinc <pap.lorinc@gmail.com>
Co-authored-by: sipa <pieter@wuille.net>